refactor(frontend): align entities/*/api/ files with naming convention
Group A of the Apollo→TanStack migration follow-up. Every file under
entities/*/api/ now ends in -from-api.ts, .query.ts, or .mutation.ts.

- Consolidate schema mutation constants (DROPDOWN_*/ENUM_*) into their
  add/remove-from-api files; delete the standalone dropdown.ts/enum.ts.
- Consolidate CREATE_PROPOSED_CHANGE into create-proposed-change-from-api.ts.
- Rename updateProposedChangeReviewFromApi.ts (kebab-case).
- Move generateRelationshipListQuery + getRelationshipParent helpers under
  nodes/relationships/api/ and suffix them .query.ts (they are query
  string builders, not API callers).
- Rename navigation/api/search.ts → search-from-api.ts.
- Rename nodes/object/api/get-display-label.ts → get-display-label-from-api.ts.
- Flatten nodes/hierarchy/api/query/ subfolder into nodes/hierarchy/api/.

No behaviour change. Verification:
  find src/entities -type f -path '*/api/*.ts' \
    ! -name '*-from-api.ts' ! -name '*.query.ts' ! -name '*.mutation.ts'
returns empty.
