+++ frontend/app/src/entities/navigation/api/search-from-api.ts
