+++ frontend/app/src/entities/nodes/hierarchy/api/get-object-ancestors.query.ts
