+++ frontend/app/src/entities/nodes/object/api/get-display-label-from-api.ts
