+++ frontend/app/src/entities/nodes/relationships/api/generate-relationship-list.query.ts
