+++ frontend/app/src/entities/nodes/relationships/api/get-relationship-parent.query.ts
