+++ frontend/app/src/entities/proposed-changes/api/create-proposed-change-from-api.ts
@@ -1,8 +1,34 @@
-import type { VariablesOf } from "gql.tada";
+import { type VariablesOf, graphql } from "gql.tada";
 
 import graphqlClient from "@/shared/api/graphql/graphqlClientApollo";
 
-import { CREATE_PROPOSED_CHANGE } from "@/entities/proposed-changes/api/createProposedChange";
+export const CREATE_PROPOSED_CHANGE = graphql(`
+  mutation CoreProposedChangeCreate(
+    $name: String!
+    $isDraft: Boolean
+    $description: String
+    $source_branch: String!
+    $destination_branch: String!
+    $reviewers: [RelatedNodeInput!]
+  ) {
+    CoreProposedChangeCreate(
+      data: {
+        name: { value: $name }
+        is_draft: { value: $isDraft }
+        description: { value: $description }
+        source_branch: { value: $source_branch }
+        destination_branch: { value: $destination_branch }
+        reviewers: $reviewers
+      }
+    ) {
+      object {
+        id
+        display_label
+      }
+      ok
+    }
+  }
+`);
 
 export interface CreateProposedChangeFromApiParams
   extends VariablesOf<typeof CREATE_PROPOSED_CHANGE> {}
